+++ ProcessHacker/kphlib/include/kphapi.h
@@ -34,6 +34,33 @@ typedef KPH_PROCESS_STATE* PKPH_PROCESS_STATE;
 #define KPH_PROCESS_NO_FILE_TRANSACTION                  0x00000080ul
 #define KPH_PROCESS_NOT_BEING_DEBUGGED                   0x00000100ul
 #define KPH_PROCESS_NO_WRITABLE_FILE_OBJECT              0x00000200ul
+#ifdef IS_KTE
+#define KPH_PROCESS_NO_VERIFY_TIMEOUT                    0x00001000ul
+
+#define KPH_PROCESS_STATE_MAXIMUM (KPH_PROCESS_SECURELY_CREATED               |\
+                                   KPH_PROCESS_VERIFIED_PROCESS               |\
+                                   KPH_PROCESS_PROTECTED_PROCESS              |\
+                                   KPH_PROCESS_NO_UNTRUSTED_IMAGES            |\
+                                   KPH_PROCESS_HAS_FILE_OBJECT                |\
+                                   KPH_PROCESS_HAS_SECTION_OBJECT_POINTERS    |\
+                                   KPH_PROCESS_NO_USER_WRITABLE_REFERENCES    |\
+                                   KPH_PROCESS_NO_FILE_TRANSACTION            |\
+                                   KPH_PROCESS_NOT_BEING_DEBUGGED             |\
+                                   KPH_PROCESS_NO_VERIFY_TIMEOUT              |\
+                                   KPH_PROCESS_NO_WRITABLE_FILE_OBJECT)
+
+#define KPH_PROCESS_STATE_HIGH    (KPH_PROCESS_VERIFIED_PROCESS               |\
+                                   KPH_PROCESS_PROTECTED_PROCESS              |\
+                                   KPH_PROCESS_NO_UNTRUSTED_IMAGES            |\
+                                   KPH_PROCESS_HAS_FILE_OBJECT                |\
+                                   KPH_PROCESS_HAS_SECTION_OBJECT_POINTERS    |\
+                                   KPH_PROCESS_NO_USER_WRITABLE_REFERENCES    |\
+                                   KPH_PROCESS_NO_FILE_TRANSACTION            |\
+                                   KPH_PROCESS_NOT_BEING_DEBUGGED             |\
+                                   KPH_PROCESS_NO_VERIFY_TIMEOUT              |\
+                                   KPH_PROCESS_NO_WRITABLE_FILE_OBJECT)
+
+#else
 
 #define KPH_PROCESS_STATE_MAXIMUM (KPH_PROCESS_SECURELY_CREATED               |\
                                    KPH_PROCESS_VERIFIED_PROCESS               |\
@@ -55,6 +82,8 @@ typedef KPH_PROCESS_STATE* PKPH_PROCESS_STATE;
                                    KPH_PROCESS_NO_FILE_TRANSACTION            |\
                                    KPH_PROCESS_NOT_BEING_DEBUGGED             |\
                                    KPH_PROCESS_NO_WRITABLE_FILE_OBJECT)
+
+#endif
 
 #define KPH_PROCESS_STATE_MEDIUM  (KPH_PROCESS_VERIFIED_PROCESS               |\
                                    KPH_PROCESS_PROTECTED_PROCESS              |\