+++ TaskExplorer/GUI/SettingsWindow.cpp
@@ -3,7 +3,9 @@
 #include "TaskExplorer.h"
 #include "../../MiscHelpers/Common/Settings.h"
 #include "../../MiscHelpers/Archive/ArchiveFS.h"
+#include "OnlineUpdater.h"
 #include <QFontDialog>
+#include <QDesktopServices>
 
 int CSettingsWindow__Chk2Int(Qt::CheckState state)
 {
@@ -40,6 +42,7 @@ CSettingsWindow::CSettingsWindow(QWidget *parent)
 	ui.tabWidget->setTabIcon(1, QIcon(":/Actions/MiscOptions"));
 	ui.tabWidget->setTabIcon(2, QIcon(":/Actions/GUI"));
 	ui.tabWidget->setTabIcon(3, QIcon(":/Actions/Settings"));
+	ui.tabWidget->setTabIcon(4, QIcon(":/Actions/Support"));
 
 	ui.tabWidget->setCurrentIndex(0);
 
@@ -108,6 +111,47 @@ CSettingsWindow::CSettingsWindow(QWidget *parent)
 	connect(ui.btnResetUiFont, SIGNAL(clicked(bool)), this, SLOT(OnResetUiFont()));
 	ui.lblUiFont->setText(QApplication::font().family());
 
+	// Updater
+	ui.cmbInterval->addItem(tr("Every Day"), 1 * 24 * 60 * 60);
+	ui.cmbInterval->addItem(tr("Every Week"), 7 * 24 * 60 * 60);
+	ui.cmbInterval->addItem(tr("Every 2 Weeks"), 14 * 24 * 60 * 60);
+	ui.cmbInterval->addItem(tr("Every 30 days"), 30 * 24 * 60 * 60);
+
+	ui.cmbUpdate->addItem(tr("Ignore"), "ignore");
+	ui.cmbUpdate->addItem(tr("Notify"), "notify");
+	ui.cmbUpdate->addItem(tr("Download & Notify"), "download");
+	ui.cmbUpdate->addItem(tr("Download & Install"), "install");
+
+	ui.cmbRelease->addItem(tr("Notify"), "notify");
+	ui.cmbRelease->addItem(tr("Download & Notify"), "download");
+	ui.cmbRelease->addItem(tr("Download & Install"), "install");
+
+	ui.chkAutoUpdate->setCheckState(CSettingsWindow__Int2Chk(theConf->GetInt("Options/CheckForUpdates", 2)));
+
+	int UpdateInterval = theConf->GetInt("Options/UpdateInterval", UPDATE_INTERVAL);
+	int pos = ui.cmbInterval->findData(UpdateInterval);
+	if (pos == -1)
+		ui.cmbInterval->setCurrentText(QString::number(UpdateInterval));
+	else
+		ui.cmbInterval->setCurrentIndex(pos);
+
+	QString ReleaseChannel = theConf->GetString("Options/ReleaseChannel", "stable");
+	ui.radStable->setChecked(ReleaseChannel == "stable");
+	ui.radPreview->setChecked(ReleaseChannel == "preview");
+
+	UpdateUpdater();
+
+	ui.cmbUpdate->setCurrentIndex(ui.cmbUpdate->findData(theConf->GetString("Options/OnNewUpdate", "ignore")));
+	ui.cmbRelease->setCurrentIndex(ui.cmbRelease->findData(theConf->GetString("Options/OnNewRelease", "download")));
+
+	connect(ui.lblCurrent, SIGNAL(linkActivated(const QString&)), this, SLOT(OnUpdate(const QString&)));
+	connect(ui.lblStable, SIGNAL(linkActivated(const QString&)), this, SLOT(OnUpdate(const QString&)));
+	connect(ui.lblPreview, SIGNAL(linkActivated(const QString&)), this, SLOT(OnUpdate(const QString&)));
+
+	connect(ui.chkAutoUpdate, SIGNAL(toggled(bool)), this, SLOT(UpdateUpdater()));
+	connect(ui.radStable, SIGNAL(toggled(bool)), this, SLOT(UpdateUpdater()));
+	connect(ui.radPreview, SIGNAL(toggled(bool)), this, SLOT(UpdateUpdater()));
+	//
 
 	ui.highlightCount->setValue(theConf->GetInt("Options/HighLoadHighlightCount", 5));
 
@@ -201,6 +245,8 @@ CSettingsWindow::CSettingsWindow(QWidget *parent)
 	connect(ui.buttonBox->button(QDialogButtonBox::Ok), SIGNAL(pressed()), this, SLOT(accept()));
 	connect(ui.buttonBox->button(QDialogButtonBox::Apply), SIGNAL(pressed()), this, SLOT(apply()));
 	connect(ui.buttonBox, SIGNAL(rejected()), this, SLOT(reject()));
+
+	connect(ui.tabWidget, SIGNAL(currentChanged(int)), this, SLOT(OnTab()));
 
 	restoreGeometry(theConf->GetBlob("SettingsWindow/Window_Geometry"));
 
@@ -301,12 +347,33 @@ void CSettingsWindow::apply()
 			theConf->SetValue("Options/ShowGrid", pItem->checkState() == Qt::Checked);
 		else if (Name != "Background"
 		 && Name != "GraphBack" && Name != "GraphFront"
-		 && Name != "PlotBack" && Name != "PlotFront" && Name != "PlotGrid" 
+		 && Name != "PlotBack" && Name != "PlotFront" && Name != "PlotGrid"
 		)
 			ColorStr += ";" + QString((pItem->checkState() == Qt::Checked) ? "true" : "false");
 
 		theConf->SetValue("Colors/" + Name, ColorStr);
 	}
+
+	// Updater
+	theConf->SetValue("Options/CheckForUpdates", CSettingsWindow__Chk2Int(ui.chkAutoUpdate->checkState()));
+
+	int UpdateInterval = ui.cmbInterval->currentData().toInt();
+	if (!UpdateInterval)
+		UpdateInterval = ui.cmbInterval->currentText().toInt();
+	if (!UpdateInterval)
+		UpdateInterval = UPDATE_INTERVAL;
+	theConf->SetValue("Options/UpdateInterval", UpdateInterval);
+
+	QString ReleaseChannel;
+	if (ui.radStable->isChecked())
+		ReleaseChannel = "stable";
+	else if (ui.radPreview->isChecked())
+		ReleaseChannel = "preview";
+	if(!ReleaseChannel.isEmpty()) theConf->SetValue("Options/ReleaseChannel", ReleaseChannel);
+
+	theConf->SetValue("Options/OnNewUpdate", ui.cmbUpdate->currentData());
+	theConf->SetValue("Options/OnNewRelease", ui.cmbRelease->currentData());
+	//
 
 	emit OptionsChanged();
 }
@@ -356,4 +423,87 @@ void CSettingsWindow::OnResetUiFont()
 {
 	QFont defaultFont = QFontDatabase::systemFont(QFontDatabase::GeneralFont);
 	ui.lblUiFont->setText(defaultFont.family());
+}
+
+void CSettingsWindow::UpdateUpdater()
+{
+	if (!ui.chkAutoUpdate->isChecked())
+	{
+		ui.cmbInterval->setEnabled(false);
+		ui.cmbUpdate->setEnabled(false);
+		ui.cmbRelease->setEnabled(false);
+		ui.lblRevision->setText(QString());
+		ui.lblRelease->setText(QString());
+	}
+	else
+	{
+		ui.cmbInterval->setEnabled(true);
+		ui.cmbUpdate->setEnabled(true);
+		ui.cmbRelease->setEnabled(true);
+
+		ui.lblRevision->setText(QString());
+		ui.lblRelease->setText(QString());
+	}
+}
+
+void CSettingsWindow::OnTab()
+{
+	// Check if we're on the updater tab (index 4)
+	if (ui.tabWidget->currentIndex() == 4)
+	{
+		if (ui.lblCurrent->text().isEmpty()) {
+			if (ui.chkAutoUpdate->checkState() == Qt::Checked)
+				GetUpdates();
+			else
+				ui.lblCurrent->setText(tr("<a href=\"check\">Check Now</a>"));
+		}
+	}
+}
+
+void CSettingsWindow::GetUpdates()
+{
+	QVariantMap Params;
+	Params["channel"] = "all";
+	theGUI->GetOnlineUpdater()->GetUpdates(this, SLOT(OnUpdateData(const QVariantMap&, const QVariantMap&)), Params);
+}
+
+QString CSettingsWindow__MkVersion(const QString& Name, const QVariantMap& Releases)
+{
+	QVariantMap Release = Releases[Name].toMap();
+	QString Version = Release.value("version").toString();
+	int iUpdate = Release["update"].toInt();
+	if(iUpdate) Version += QChar('a' + (iUpdate - 1));
+	return QString("<a href=\"%1\">%2</a>").arg(Name, Version);
+}
+
+void CSettingsWindow::OnUpdateData(const QVariantMap& Data, const QVariantMap& Params)
+{
+	if (Data.isEmpty() || Data["error"].toBool())
+		return;
+
+	m_UpdateData = Data;
+	QVariantMap Releases = m_UpdateData["releases"].toMap();
+	ui.lblCurrent->setText(tr("%1 (Current)").arg(COnlineUpdater::GetCurrentVersion()));
+	ui.lblStable->setText(CSettingsWindow__MkVersion("stable", Releases));
+	ui.lblPreview->setText(CSettingsWindow__MkVersion("preview", Releases));
+}
+
+void CSettingsWindow::OnUpdate(const QString& Channel)
+{
+	if (Channel == "check") {
+		GetUpdates();
+		return;
+	}
+
+	QVariantMap Releases = m_UpdateData["releases"].toMap();
+	QVariantMap Release = Releases[Channel].toMap();
+
+	QString VersionStr = Release["version"].toString();
+	if (VersionStr.isEmpty())
+		return;
+
+	QString InfoUrl = Release["infoUrl"].toString();
+	if (InfoUrl.isEmpty())
+		InfoUrl = "https://xanasoft.com/go.php?to=sbie-get";
+	QDesktopServices::openUrl(InfoUrl);
 }