+++ TaskExplorer/GUI/SettingsWindow.h
@@ -26,8 +26,17 @@ private slots:
 	void OnSelectUiFont();
 	void OnResetUiFont();
 
+	void GetUpdates();
+	void OnUpdateData(const QVariantMap& Data, const QVariantMap& Params);
+	void OnUpdate(const QString& Channel);
+	void UpdateUpdater();
+
+	void OnTab();
+
 protected:
 	void closeEvent(QCloseEvent *e);
+
+	QVariantMap m_UpdateData;
 
 private:
 	Ui::SettingsWindow ui;